+++ apps/wallet/src/ui/app/pages/home/index.tsx
@@ -31,7 +31,6 @@ const HomePage = ({ disableNavigation }: HomePageProps) => {
 
 export default HomePage;
 export { default as NftsPage } from './nfts';
-export { default as HiddenAssetsPage } from './hidden-assets';
 export { default as AssetsPage } from './assets';
 export { default as TokensPage } from './tokens';
 export { default as TransactionBlocksPage } from './transactions';