+++ apps/wallet/src/ui/app/pages/home/nfts/NonVisualAssets.tsx
@@ -6,6 +6,7 @@ import { ExplorerLink, ExplorerLinkType } from '_components';
 import { type IotaObjectData } from '@iota/iota-sdk/client';
 import { formatAddress, parseStructTag } from '@iota/iota-sdk/utils';
 import { Card, CardAction, CardActionType, CardBody, CardType } from '@iota/apps-ui-kit';
+import { ArrowTopRight } from '@iota/ui-icons';
 
 interface NonVisualAssetsProps {
     items: IotaObjectData[];
@@ -30,7 +31,10 @@ export default function NonVisualAssets({ items }: NonVisualAssetsProps) {
                                         title={formatAddress(item.objectId!)}
                                         subtitle={`${formatAddress(address)}::${module}::${name}`}
                                     />
-                                    <CardAction type={CardActionType.Link} isExternalLink />
+                                    <CardAction
+                                        type={CardActionType.Link}
+                                        icon={<ArrowTopRight />}
+                                    />
                                 </Card>
                             </ExplorerLink>
                         );