+++ apps/wallet/src/ui/app/pages/home/transfer-coin/PreviewTransfer.tsx
@@ -2,11 +2,22 @@
 // Modifications Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import { Text } from '_app/shared/text';
-import { TxnAddress, TxnAmount } from '_components';
+import { CoinIcon } from '_components';
 import { useActiveAddress } from '_src/ui/app/hooks/useActiveAddress';
 import { GAS_SYMBOL } from '_src/ui/app/redux/slices/iota-objects/Coin';
-import { parseAmount, useCoinMetadata } from '@iota/core';
+import { parseAmount, useCoinMetadata, useFormatCoin } from '@iota/core';
+import {
+    Card,
+    CardAction,
+    CardActionType,
+    CardBody,
+    CardImage,
+    CardType,
+    Divider,
+    ImageType,
+    KeyValueInfo,
+} from '@iota/apps-ui-kit';
+import { useAddressLink } from '_app/hooks/useAddressLink';
 
 export type PreviewTransferProps = {
     coinType: string;
@@ -26,26 +37,42 @@ export function PreviewTransfer({
     const accountAddress = useActiveAddress();
     const { data: metadata } = useCoinMetadata(coinType);
     const amountWithoutDecimals = parseAmount(amount, metadata?.decimals ?? 0);
+    const [formatted, symbol] = useFormatCoin(
+        Math.abs(Number(amountWithoutDecimals.toString())),
+        coinType,
+    );
+
+    const addrFrom = useAddressLink(accountAddress);
+    const addrTo = useAddressLink(to);
 
     return (
-        <div className="divide-steel/20 flex w-full flex-col divide-x-0 divide-y divide-solid px-2.5">
-            <TxnAmount
-                amount={amountWithoutDecimals.toString()}
-                label="Sending"
-                coinType={coinType}
-                approximation={approximation}
-            />
-            <TxnAddress address={accountAddress || ''} label="From" />
-            <TxnAddress address={to} label="To" />
-            <div className="mb-5 flex w-full justify-between gap-2 pt-3.5">
-                <div className="flex gap-1">
-                    <Text variant="body" color="gray-80" weight="medium">
-                        Estimated Gas Fees
-                    </Text>
-                </div>
-                <Text variant="body" color="gray-90" weight="medium">
-                    {gasBudget} {GAS_SYMBOL}
-                </Text>
+        <div className="flex w-full flex-col gap-md">
+            <Card type={CardType.Filled}>
+                <CardImage type={ImageType.BgSolid}>
+                    <div className="flex h-10 w-10 items-center justify-center rounded-full border border-shader-neutral-light-8  text-neutral-10">
+                        <CoinIcon coinType={coinType} />
+                    </div>
+                </CardImage>
+                <CardBody
+                    title={`${approximation ? '~' : ''}${formatted} ${symbol}`}
+                    subtitle="Amount"
+                />
+                <CardAction type={CardActionType.SupportingText} />
+            </Card>
+            <div className="flex flex-col gap-md--rs p-sm--rs">
+                <KeyValueInfo
+                    keyText={'From'}
+                    valueText={addrFrom.address}
+                    valueLink={addrFrom.explorerHref}
+                />
+                <Divider />
+                <KeyValueInfo
+                    keyText={'To'}
+                    valueText={addrTo.address}
+                    valueLink={addrTo.explorerHref}
+                />
+                <Divider />
+                <KeyValueInfo keyText={'Est. Gas Fees'} valueText={`${gasBudget} ${GAS_SYMBOL}`} />
             </div>
         </div>
     );