+++ crates/iota-sdk/examples/event_api.rs
@@ -17,17 +17,17 @@ use utils::{setup_for_write, split_coin_digest};
 
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {
-    let (iota, active_address, _second_address) = setup_for_write().await?;
+    let (client, active_address, _second_address) = setup_for_write().await?;
 
     println!(" *** Get events *** ");
     // for demonstration purposes, we set to make a transaction
-    let digest = split_coin_digest(&iota, &active_address).await?;
-    let events = iota.event_api().get_events(digest).await?;
+    let digest = split_coin_digest(&client, &active_address).await?;
+    let events = client.event_api().get_events(digest).await?;
     println!("{:?}", events);
     println!(" *** Get events ***\n ");
 
     let descending = true;
-    let query_events = iota
+    let query_events = client
         .event_api()
         .query_events(EventFilter::All(vec![]), None, Some(5), descending) // query first 5 events in descending order
         .await?;