+++ crates/iota-sdk/examples/read_api.rs
@@ -12,22 +12,21 @@
 //! cargo run --example read_api
 
 mod utils;
+
 use iota_sdk::{
-    rpc_types::{
-        IotaGetPastObjectRequest, IotaObjectDataOptions, IotaTransactionBlockResponseOptions,
-    },
+    rpc_types::{IotaGetPastObjectRequest, IotaObjectDataOptions},
     types::base_types::ObjectID,
 };
-use utils::{setup_for_write, split_coin_digest};
+use utils::setup_for_write;
 
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {
-    let (iota, active_address, _) = setup_for_write().await?;
+    let (client, active_address, _) = setup_for_write().await?;
 
     // ************ READ API ************ //
     println!("// ************ READ API ************ //\n");
     // Owned Objects
-    let owned_objects = iota
+    let owned_objects = client
         .read_api()
         .get_owned_objects(active_address, None, None, Some(5))
         .await?;
@@ -37,7 +36,7 @@ async fn main() -> Result<(), anyhow::Error> {
 
     // Dynamic Fields
     let parent_object_id = ObjectID::from_address(active_address.into());
-    let dynamic_fields = iota
+    let dynamic_fields = client
         .read_api()
         .get_dynamic_fields(parent_object_id, None, None)
         .await?;
@@ -46,7 +45,7 @@ async fn main() -> Result<(), anyhow::Error> {
     println!(" *** Dynamic Fields ***\n");
     if let Some(dynamic_field_info) = dynamic_fields.data.into_iter().next() {
         println!(" *** First Dynamic Field ***");
-        let dynamic_field = iota
+        let dynamic_field = client
             .read_api()
             .get_dynamic_field_object(parent_object_id, dynamic_field_info.name)
             .await?;
@@ -75,7 +74,7 @@ async fn main() -> Result<(), anyhow::Error> {
         show_storage_rebate: true,
     };
 
-    let past_object = iota
+    let past_object = client
         .read_api()
         .try_get_parsed_past_object(object_id, version, iota_data_options.clone())
         .await?;
@@ -84,7 +83,7 @@ async fn main() -> Result<(), anyhow::Error> {
     println!(" *** Past Object ***\n");
 
     let iota_get_past_object_request = past_object.clone().into_object()?;
-    let multi_past_object = iota
+    let multi_past_object = client
         .read_api()
         .try_multi_get_parsed_past_object(
             vec![IotaGetPastObjectRequest {
@@ -99,7 +98,7 @@ async fn main() -> Result<(), anyhow::Error> {
     println!(" *** Multi Past Object ***\n");
 
     // Object with options
-    let object_with_options = iota
+    let object_with_options = client
         .read_api()
         .get_object_with_options(iota_get_past_object_request.object_id, iota_data_options)
         .await?;
@@ -109,36 +108,14 @@ async fn main() -> Result<(), anyhow::Error> {
     println!(" *** Object with Options ***\n");
 
     println!(" *** Chain identifier *** ");
-    println!("{:?}", iota.read_api().get_chain_identifier().await?);
+    println!("{:?}", client.read_api().get_chain_identifier().await?);
     println!(" *** Chain identifier ***\n ");
 
     println!(" *** Protocol Config *** ");
-    println!("{:?}", iota.read_api().get_protocol_config(None).await?);
+    println!("{:?}", client.read_api().get_protocol_config(None).await?);
     println!(" *** Protocol Config ***\n ");
 
-    // we make a dummy transaction which returns a transaction digest
-    let tx_digest = split_coin_digest(&iota, &active_address).await?;
-    println!(" *** Transaction data *** ");
-    let tx_response = iota
-        .read_api()
-        .get_transaction_with_options(
-            tx_digest,
-            IotaTransactionBlockResponseOptions {
-                show_input: true,
-                show_raw_input: true,
-                show_effects: true,
-                show_events: true,
-                show_object_changes: true,
-                show_balance_changes: true,
-                show_raw_effects: true,
-            },
-        )
-        .await?;
-    println!("Transaction succeeded: {:?}\n\n", tx_response.status_ok());
-
-    println!("Transaction data: {tx_response:?}");
-
-    let tx_blocks = iota.read_api().get_total_transaction_blocks().await?;
+    let tx_blocks = client.read_api().get_total_transaction_blocks().await?;
     println!("Total transaction blocks {tx_blocks}");
     // ************ END OF READ API ************ //