+++ crates/iota-sdk/examples/sign_tx_guide.rs
@@ -37,7 +37,7 @@ use utils::request_tokens_from_faucet;
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {
     // set up iota client for the desired network.
-    let iota_client = IotaClientBuilder::default().build_testnet().await?;
+    let client = IotaClientBuilder::default().build_testnet().await?;
 
     // deterministically generate a keypair, testing only, do not use for mainnet,
     // use the next section to randomly generate a keypair instead.
@@ -102,8 +102,8 @@ async fn main() -> Result<(), anyhow::Error> {
     println!("Sender: {sender:?}");
 
     // make sure the sender has a gas coin as an example.
-    request_tokens_from_faucet(sender, &iota_client).await?;
-    let gas_coin = iota_client
+    request_tokens_from_faucet(sender, &client).await?;
+    let gas_coin = client
         .coin_read_api()
         .get_coins(sender, None, None, None)
         .await?
@@ -120,7 +120,7 @@ async fn main() -> Result<(), anyhow::Error> {
     };
 
     let gas_budget = 5_000_000;
-    let gas_price = iota_client.read_api().get_reference_gas_price().await?;
+    let gas_price = client.read_api().get_reference_gas_price().await?;
 
     // create the transaction data that will be sent to the network.
     let tx_data = TransactionData::new_programmable(
@@ -153,7 +153,7 @@ async fn main() -> Result<(), anyhow::Error> {
     assert!(res.is_ok());
 
     // execute the transaction.
-    let transaction_response = iota_client
+    let transaction_response = client
         .quorum_driver_api()
         .execute_transaction_block(
             iota_types::transaction::Transaction::from_generic_sig_data(